circulating supply error fixed

+++ crypto_app/Model/Coin.swift
@@ -15,7 +15,7 @@ struct Coin: Codable {
     let priceChange24H, priceChangePercentage24H: Double
     let marketCapChange24H: Int?
     let marketCapChangePercentage24H: Double?
-    let circulatingSupply, totalSupply, maxSupply, ath: Int?
+    let circulatingSupply, totalSupply, maxSupply, ath: Double?
     let athChangePercentage: Double?
     let athDate: String?
     let atl, atlChangePercentage: Double?
